# backend/openapi.yaml
openapi: 3.0.3
info:
  title: USD to LKR Exchange Rate API
  version: 2.1.0
  description: |
    API to fetch current and historical exchange rates from USD to LKR from multiple banks.
    
    ## Currently Supported Banks
    - SAMPATH - Sampath Bank
    
    ## Planned Banks (Coming Soon)
    - NSB - National Savings Bank
    - SEYLAN - Seylan Bank
    - NATION - Nations Trust Bank
    - COMMERCIAL - Commercial Bank
    - HNB - Hatton National Bank
    
    ## Features
    - Real-time rate fetching from multiple banks
    - Historical data with period filtering (week, month, year)
    - Bank-specific data filtering
    - Health monitoring endpoint
    
    ## Bank Parameter
    Most endpoints support an optional `bank` query parameter to filter results by a specific bank.
    If not provided, data from all available banks will be returned.
    
    ## Period Parameter
    History endpoint supports `period` parameter to filter by time range:
    - `week`: Last 7 days (daily data points)
    - `month`: Last 30 days (daily data points)  
    - `year`: Last 365 days (weekly data points)

servers:
  - url: http://localhost:8080
    description: Local development server


components:
  schemas:
    ExchangeRate:
      type: object
      required:
        - rate
        - fetchedAt
        - bank
      properties:
        rate:
          type: number
          format: float
          description: Exchange rate from USD to LKR
          example: 326.45
        fetchedAt:
          type: string
          format: date-time
          description: Timestamp when the rate was fetched
          example: "2025-07-10T14:30:00Z"
        bank:
          type: string
          description: Bank name (constant)
          enum: [SAMPATH, COMMERCIAL, HNB, NSB, SEYLAN, NATION]
          example: "SAMPATH"
    
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Bank not supported: INVALID"
        details:
          type: array
          items:
            type: string
          description: Additional error details
    
    FetchRatesResponse:
      type: object
      properties:
        rates:
          type: array
          items:
            $ref: '#/components/schemas/ExchangeRate'
        warnings:
          type: array
          items:
            type: string
          description: Warnings for failed bank extractions (only present if some banks failed)
    
    BanksResponse:
      type: object
      properties:
        banks:
          type: array
          items:
            type: string
            enum: [SAMPATH, COMMERCIAL, HNB, NSB, SEYLAN, NATION]
          description: List of supported bank names
          example: ["SAMPATH", "COMMERCIAL", "HNB"]
    
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          description: Server health status
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          description: Current server timestamp
          example: "2025-07-11T01:16:13.140775+05:30"
        banks:
          type: integer
          description: Number of configured bank extractors
          example: 3

  parameters:
    BankParameter:
      name: bank
      in: query
      required: false
      description: |
        Optional bank name to filter results. 
        Must be one of the supported bank constants: SAMPATH, COMMERCIAL, HNB, NSB, SEYLAN, NATION
      schema:
        type: string
        enum: [SAMPATH, COMMERCIAL, HNB, NSB, SEYLAN, NATION]
      example: "SAMPATH"
    
    PeriodParameter:
      name: period
      in: query
      required: false
      description: |
        Time period for historical data filtering.
        - `week`: Last 7 days (7 daily data points)
        - `month`: Last 30 days (30 daily data points)
        - `year`: Last 365 days (52 weekly data points)
      schema:
        type: string
        enum: [week, month, year]
        default: month
      example: "month"

paths:
  /fetch-rate:
    get:
      summary: Fetch current USD to LKR exchange rate
      description: |
        Fetches current exchange rates from banks and stores them in the database.
        
        - If `bank` parameter is provided, fetches from that specific bank only
        - If no `bank` parameter, fetches from all available banks
        - Returns single rate object for specific bank, array for multiple banks
      parameters:
        - $ref: '#/components/parameters/BankParameter'
      responses:
        '200':
          description: Successful response with current exchange rate(s)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExchangeRate'
                  - $ref: '#/components/schemas/FetchRatesResponse'
              examples:
                single_bank:
                  summary: Response for specific bank
                  value:
                    rate: 326.45
                    fetchedAt: "2025-07-10T14:30:00Z"
                    bank: "SAMPATH"
                multiple_banks:
                  summary: Response for all banks
                  value:
                    rates:
                      - rate: 326.45
                        fetchedAt: "2025-07-10T14:30:00Z"
                        bank: "SAMPATH"
                      - rate: 325.80
                        fetchedAt: "2025-07-10T14:30:15Z"
                        bank: "COMMERCIAL"
                    warnings:
                      - "Error fetching rate from HNB: connection timeout"
        '400':
          description: Bad request - invalid bank parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Bank not supported: INVALID"
        '500':
          description: Internal server error - failed to fetch rates from any bank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Failed to fetch rates from any bank"
                details:
                  - "Error fetching rate from SAMPATH: API unavailable"
                  - "Error inserting rate for COMMERCIAL to DB: connection failed"
  /latest-rate:
    get:
      summary: Get latest USD to LKR exchange rate
      description: |
        Returns the latest available exchange rate from the database.
        
        - If `bank` parameter is provided, returns latest rate for that specific bank
        - If no `bank` parameter, returns the latest rate from any bank
      parameters:
        - $ref: '#/components/parameters/BankParameter'
      responses:
        '200':
          description: Latest exchange rate response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRate'
              example:
                rate: 327.1
                fetchedAt: "2025-07-10T14:30:00Z"
                bank: "SAMPATH"
        '404':
          description: No rates found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                no_rates_for_bank:
                  summary: No rates for specific bank
                  value:
                    error: "No rates found for bank: COMMERCIAL"
                no_rates_at_all:
                  summary: No rates found at all
                  value:
                    error: "No rates found"
  /history:
    get:
      summary: Get historical USD to LKR rates
      description: |
        Returns historical exchange rates from the database with optional period filtering.
        
        - If `bank` parameter is provided, returns history for that specific bank
        - If no `bank` parameter, returns history from all banks
        - `period` parameter controls the time range and data density:
          - `week`: Last 7 days with daily data points
          - `month`: Last 30 days with daily data points (default)
          - `year`: Last 365 days with weekly data points
        - Results are sorted by fetchedAt in descending order (newest first)
      parameters:
        - $ref: '#/components/parameters/BankParameter'
        - $ref: '#/components/parameters/PeriodParameter'
      responses:
        '200':
          description: Historical rate data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExchangeRate'
              examples:
                week_data:
                  summary: Weekly data for specific bank
                  value:
                    - rate: 327.1
                      fetchedAt: "2025-07-11T14:30:00Z"
                      bank: "SAMPATH"
                    - rate: 326.8
                      fetchedAt: "2025-07-10T14:30:00Z"
                      bank: "SAMPATH"
                    - rate: 326.5
                      fetchedAt: "2025-07-09T14:30:00Z"
                      bank: "SAMPATH"
                month_data:
                  summary: Monthly data for specific bank
                  value:
                    - rate: 327.1
                      fetchedAt: "2025-07-11T14:30:00Z"
                      bank: "COMMERCIAL"
                    - rate: 326.8
                      fetchedAt: "2025-07-10T14:30:00Z"
                      bank: "COMMERCIAL"
                year_data:
                  summary: Yearly data (weekly points)
                  value:
                    - rate: 327.1
                      fetchedAt: "2025-07-11T14:30:00Z"
                      bank: "HNB"
                    - rate: 325.8
                      fetchedAt: "2025-07-04T14:30:00Z"
                      bank: "HNB"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Failed to fetch history"

  /banks:
    get:
      summary: Get list of supported banks
      description: |
        Returns a list of all supported bank names (constants).
        Useful for frontend applications to populate bank selection dropdowns.
      responses:
        '200':
          description: List of supported banks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BanksResponse'
              example:
                banks: ["SAMPATH", "COMMERCIAL", "HNB"]

  /health:
    get:
      summary: Health check endpoint
      description: |
        Returns server health status and basic information.
        Useful for monitoring and debugging.
      responses:
        '200':
          description: Server health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: "healthy"
                timestamp: "2025-07-11T01:16:13.140775+05:30"
                banks: 3